lwip: add pingpong run config with client and server connected via nic_bridge
The run script reads the server's IP address from /tmp/test-ping_server

Remark: There is a race between the client and server.
Whichever opens a session to the nic_bridge first gets the first
(virtual) MAC address. Therefore the IP addresses acquired via
DHCP depend on the order in which the client and server
open their sessions to the nic_bridge.